move cognitoId to studentId responsibility
