+++ lambdas/orchestrators/WorkerDispatcher/package.json
@@ -1,13 +1,14 @@
 {
-  "name": "worker-dispatcher",
+  "name": "student-query-worker-dispatcher",
   "version": "1.0.0",
-  "description": "Lambda function to dispatch tasks to worker Lambdas",
+  "description": "Lambda function that dispatches events to worker lambdas based on LLM analysis",
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "dependencies": {
-    "aws-sdk": "^2.1531.0"
+    "aws-sdk": "^2.1450.0",
+    "pg": "^8.11.3"
   },
   "engines": {
     "node": ">=16.x"
